[tests] move wait_until() to utils.py
